Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
sqlalchemy
pg8000 
langchain-google-genai
google-generativeai
langchain-core
langchain
langchain-community
langsmith
Pillow
pillow-heif
dropbox
cloud-sql-python-connector
google-cloud-storage